Added php version
