+++ nodejs/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": " nodemon server.js"
+    "start": " nodemon src/server.js"
   },
   "repository": {
     "type": "git",
