+++ nodejs/src/uyga-vazifa.graphql
