+++ php/composer.json
@@ -1,0 +1,12 @@
+{
+    "name": "graphql/php",
+    "authors": [
+        {
+            "name": "Shoxrux Rashidov",
+            "email": "dev.rashidov@gmail.com"
+        }
+    ],
+    "require": {
+        "webonyx/graphql-php": "^14.11"
+    }
+}
